fix: upgrade to csi-provisioner v5.0.2 on v4.8.0 chart

+++ pkg/nfs/controllerserver.go
@@ -255,8 +255,12 @@ func (cs *ControllerServer) DeleteVolume(ctx context.Context, req *csi.DeleteVol
 
 			// archive subdirectory under base-dir, remove stale archived copy if exists.
 			klog.V(2).Infof("archiving subdirectory %s --> %s", internalVolumePath, archivedInternalVolumePath)
-			if err = os.RemoveAll(archivedInternalVolumePath); err != nil {
-				return nil, status.Errorf(codes.Internal, "failed to delete archived subdirectory %s: %v", archivedInternalVolumePath, err.Error())
+			if cs.Driver.removeArchivedVolumePath {
+				klog.V(2).Infof("removing archived subdirectory at %v", archivedInternalVolumePath)
+				if err = os.RemoveAll(archivedInternalVolumePath); err != nil {
+					return nil, status.Errorf(codes.Internal, "failed to delete archived subdirectory %s: %v", archivedInternalVolumePath, err.Error())
+				}
+				klog.V(2).Infof("removed archived subdirectory at %v", archivedInternalVolumePath)
 			}
 			if err = os.Rename(internalVolumePath, archivedInternalVolumePath); err != nil {
 				return nil, status.Errorf(codes.Internal, "archive subdirectory(%s, %s) failed with %v", internalVolumePath, archivedInternalVolumePath, err.Error())
